move api flags to `api_common.go`
